feat: add createProduct helper. update catalogBatch lambda to execute and populate both tables
