+++ src/lambda/products/package.json
